api/requirements.txt
fastapi
gunicorn
uvicorn
requests
beautifulsoup4
readability-lxml
google-api-python-client
